+++ src/dataManaging/files/MyFile.java
@@ -5,7 +5,7 @@ import java.io.File;
 import javax.swing.JFileChooser;
 
 import dataManaging.AbstractItem;
-import tools.CancelledCommandException;
+import tools.exceptions.CancelledCommandException;
 
 public class MyFile extends AbstractItem {
