+++ src/tools/MySupplier.java
@@ -1,5 +1,6 @@
 package tools;
 
+import tools.exceptions.CancelledCommandException;
 
 public interface MySupplier<T> {
