+++ src/tools/exceptions/InterruptedFileTransferException.java
@@ -1,4 +1,4 @@
-package tools;
+package tools.exceptions;
 
 public class InterruptedFileTransferException extends Exception {
